Rename files to match proejct naming convention
