+++ public/video-ui/src/components/FormFields/richtext/utils/linkValidator.ts
